feat(docker): minimal bare-circle image + non-root world-owner runtime
`docker build -t circle-min .` now produces a MINIMAL circle by default:
no OpenTelemetry, no admin API (the latter needs the web-admin sidecar).
The combined stack is unaffected — tools/docker/docker-compose.full.yml
already passes WITH_OTEL/WITH_ADMIN_API=true explicitly.

Replace the python entrypoint with a small shell entrypoint that:
- passes its args straight to circle (a leading `--` is accepted), so a
  plain `docker run ... circle-min -d /world 4000` just works — no
  --entrypoint override, no env juggling;
- drops privileges to the OWNER of the mounted world (su-exec), or to
  PUID:PGID if set. circle no longer runs as root: reads happen as that
  uid and files written back to the world keep that owner on the host
  instead of root/a random uid. This also makes the full stack non-root;
  web-admin stays root and can still reach the socket;
- cd's into the world so circle's CWD-relative files (syslog) are
  writable by the unprivileged uid;
- forces admin_api/otel config only when FORCE_CONFIG=on (off by default
  for the minimal image; compose.full.yml sets it on).

The config-forcing logic moves out of entrypoint.py into
configure_world.py (config only, no exec), invoked by the entrypoint.
