+++ Dockerfile
@@ -1,24 +1,27 @@
 # syntax=docker/dockerfile:1
 #
 # Образ игрового сервера Bylins (circle). Собирается ИЗ ЛОКАЛЬНОГО исходника.
-# Опциональные возможности включаются build-аргументами:
-#   WITH_YAML=true       YAML-формат мира (по умолчанию включён)
-#   WITH_SQLITE=false    SQLite-формат мира (по умолчанию выключен)
-#   WITH_ADMIN_API=true  Admin API (Unix-сокет) — нужен для web-админки
-#   WITH_OTEL=true       OpenTelemetry (метрики/трейсы/логи по OTLP HTTP)
+#
+# По умолчанию собирается МИНИМАЛЬНЫЙ циркуль: без телеметрии и без admin API
+# (последний тянет за собой web-админку). Возможности включаются build-аргументами:
+#   WITH_YAML=true        YAML-формат мира (по умолчанию включён)
+#   WITH_SQLITE=false     SQLite-формат мира (по умолчанию выключен)
+#   WITH_ADMIN_API=false  Admin API (Unix-сокет) — нужен для web-админки
+#   WITH_OTEL=false       OpenTelemetry (метрики/трейсы/логи по OTLP HTTP)
 # Legacy-формат мира поддерживается всегда, отдельного флага не требует.
+# Combined-стек (circle + web-admin + телеметрия) собирается через
+# tools/docker/docker-compose.full.yml — он передаёт WITH_OTEL/ADMIN_API=true.
 #
-# Сборка:
-#   docker build -t mud-server .
-#   docker build -t mud-server --build-arg WITH_OTEL=false .   # без телеметрии (быстрее)
+# Сборка минимального образа:
+#   docker build -t circle-min .
 #
-# Запуск (turnkey: на вход — каталог мира; entrypoint форсирует admin_api+otel-only):
-#   docker run -d --network host \
-#     -v /path/to/world:/world \
-#     --name mud mud-server
-# Циркуль запускается с host-сетью: игровой порт сразу наружу, а OTLP-endpoint
-# берётся из configuration.xml мира (localhost:<порт> бьёт в коллектор этого
-# сервера). Форсирование отключается через -e FORCE_CONFIG=0.
+# Запуск (аргументы передаются циркулю как есть; порт хоста = порту циркуля):
+#   docker run -d -p 127.0.0.1:4000:4000 -v /abs/world/lib:/world circle-min -d /world 4000
+#   docker run -d -p 127.0.0.1:4000:4000 -v /abs/world/lib:/world circle-min   # дефолт: -d /world 4000
+#
+# Циркуль бежит под владельцем смонтированного мира (не root): чтение от нужного
+# uid, а файлы, записанные в мир, остаются за тем же владельцем на хосте.
+# Override владельца — PUID/PGID. Форс конфига (admin_api+otel) — при FORCE_CONFIG=on.
 
 ARG BUILD_TYPE=release
 ARG OTEL_VERSION=1.24.0
@@ -27,8 +30,8 @@ ARG OTEL_VERSION=1.24.0
 FROM alpine:3.23 AS builder
 ARG BUILD_TYPE
 ARG OTEL_VERSION
-ARG WITH_OTEL=true
-ARG WITH_ADMIN_API=true
+ARG WITH_OTEL=false
+ARG WITH_ADMIN_API=false
 ARG WITH_YAML=true
 ARG WITH_SQLITE=false
 
@@ -71,25 +74,30 @@ RUN OTEL_OPT=$([ "$WITH_OTEL" = "true" ] && echo system || echo disabled); \
 
 # ───────────────────────── Этап 2: рантайм ─────────────────────────
 FROM alpine:3.23
-ARG WITH_OTEL=true
+ARG WITH_OTEL=false
 ARG WITH_SQLITE=false
 
-# Рантайм-зависимости: динамические библиотеки циркуля + python3 для entrypoint.
+# Рантайм-зависимости: динамические библиотеки циркуля, su-exec для сброса
+# привилегий до владельца мира, python3 для форса конфига (configure_world.py).
 # protobuf/abseil подтягиваются otel-эксportёром (Alpine отдаёт их shared-либами).
-RUN apk add --no-cache libstdc++ openssl curl zlib expat python3 && \
+RUN apk add --no-cache libstdc++ openssl curl zlib expat su-exec python3 && \
     if [ "$WITH_OTEL" = "true" ]; then apk add --no-cache protobuf abseil-cpp; fi && \
     if [ "$WITH_SQLITE" = "true" ]; then apk add --no-cache sqlite-libs; fi
 
 COPY --from=builder /mud/mud/build/circle /usr/local/bin/circle
-COPY --from=builder /mud/mud/tools/docker/entrypoint.py /usr/local/bin/entrypoint.py
+COPY --from=builder /mud/mud/tools/docker/docker-entrypoint.sh /usr/local/bin/docker-entrypoint.sh
+COPY --from=builder /mud/mud/tools/docker/configure_world.py /usr/local/bin/configure_world.py
+RUN chmod +x /usr/local/bin/docker-entrypoint.sh
 
 # Значения по умолчанию (переопределяются окружением/compose).
+# FORCE_CONFIG=0: минимальный образ конфиг мира не трогает; combined-стек
+# (compose.full.yml) выставляет FORCE_CONFIG=1 сам.
 ENV WORLD_DIR=/world \
     MUD_PORT=4000 \
-    FORCE_CONFIG=1 \
+    FORCE_CONFIG=0 \
     ADMIN_SOCKET_REL=run/admin_api.sock
 
 EXPOSE 4000
 VOLUME /world
 
-ENTRYPOINT ["python3", "/usr/local/bin/entrypoint.py"]
+ENTRYPOINT ["/usr/local/bin/docker-entrypoint.sh"]
